pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "morning-briefs"
version = "0.1.0"
description = "Local-first daily eight-minute work morning briefing generator."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "beautifulsoup4>=4.12.3",
  "feedparser>=6.0.11",
  "flask>=3.0.3",
  "openai>=2.0.0",
  "python-dotenv>=1.0.1",
  "requests>=2.32.3"
]

[project.scripts]
morning-briefs = "morning_briefs.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["morning_briefs*"]

[tool.setuptools.package-data]
morning_briefs = ["web/templates/*.html", "web/static/*"]
